Thanks to the people@elua-dev list, I finally understood what's ARM EABI, so I thought I make good use of it :), so I enabled it for ALL ARM and Cortex targets alike. Now you can specify the toolchain to use like this:
$ scons board=... toolchain=arm-gcc|arm-gcc-eabi
(the default is still arm-gcc, but this can be changed by editing SConstruct).

This is a Very Good thing for us, because users won't need to build their own toolchains anymore, they can just download the one from CS and use it. Thanks again for this.
Actually, the whole build system was changed to support arbitrary toolchains (although there's much work to be done in this area).
Also, changed a few things in the STM32 port (the linker script file, the stack definitions).
And other (very minor) changes (mostly to fix some compiler warnings).

+++ src/platform/stm32/systick.c
@@ -33,7 +33,7 @@ void SysTick_Config(void)
 *******************************************************************************/
 void Delay(u32 nCount)
 {
-  printf("Delay(%d)\n", nCount);
+  printf("Delay(%u)\n", (unsigned)nCount);
   TimingDelay = nCount;
 
   /* Enable the SysTick Counter */
